New feature: Store support for Select editor
This will allow to pull Select options data from a server. Instead of
providing values for the attribute using ‘Select’ editor, a store
attributes should be provided (with url). The url can have variables
inside it that correspond to other attributes of the same object (i.e.
http://example.com/users/{id}, the id will be replaced with the value
of the ‘id’ attribute. For an example and test case, check
gform/tests/test_primitive_select_store.html